Add validations to path variable.

+++ src/main/java/com/momo/savanger/web/BudgetController.java
@@ -1,24 +1,30 @@
 package com.momo.savanger.web;
 
-import com.momo.savanger.api.budget.Budget;
 import com.momo.savanger.api.budget.BudgetMapper;
 import com.momo.savanger.api.budget.BudgetService;
+import com.momo.savanger.api.budget.constraints.CanAccessBudget;
+import com.momo.savanger.api.budget.constraints.ValidBudget;
 import com.momo.savanger.api.budget.dto.AssignParticipantDto;
 import com.momo.savanger.api.budget.dto.BudgetDto;
 import com.momo.savanger.api.budget.dto.CreateBudgetDto;
 import com.momo.savanger.api.budget.dto.UnassignParticipantDto;
 import com.momo.savanger.api.user.User;
 import com.momo.savanger.constants.Endpoints;
+import com.momo.savanger.error.ApiErrorCode;
+import com.momo.savanger.error.ApiException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+@Validated
 @RestController
 @RequiredArgsConstructor
 @PreAuthorize("isFullyAuthenticated()")
@@ -37,26 +43,31 @@ public class BudgetController {
         );
     }
 
+    @GetMapping(Endpoints.BUDGET)
+    public BudgetDto getBudget(@PathVariable("id") @CanAccessBudget Long budgetId) {
+        return this.budgetMapper.toBudgetDto(this.budgetService.findByIdFetchAll(budgetId));
+    }
+
     @PostMapping(Endpoints.PARTICIPANTS)
-    public Integer assignParticipants(@PathVariable Long id,
+    public BudgetDto assignParticipants(@PathVariable @ValidBudget Long id,
             @Valid @RequestBody AssignParticipantDto dto) {
 
+        if (!id.equals(dto.getBudgetRef().getId())) {
+            throw ApiException.with(ApiErrorCode.ERR_0001);
+        }
         this.budgetService.addParticipant(dto);
-
-        Budget budget = this.budgetService.findById(id);
-
-        return budget.getParticipants().size();
+        return this.budgetMapper.toBudgetDto(this.budgetService.findByIdFetchAll(id));
     }
 
     @DeleteMapping(Endpoints.PARTICIPANTS)
-    public Integer unassignParticipants(@PathVariable Long id,
+    public BudgetDto unassignParticipants(@PathVariable @ValidBudget Long id,
             @Valid @RequestBody UnassignParticipantDto dto) {
 
+        if (!id.equals(dto.getBudgetRef().getId())) {
+            throw ApiException.with(ApiErrorCode.ERR_0001);
+        }
         this.budgetService.deleteParticipant(dto);
-
-        Budget budget = this.budgetService.findById(id);
-
-        return budget.getParticipants().size();
+        return this.budgetMapper.toBudgetDto(this.budgetService.findByIdFetchAll(id));
     }
